requirements.txt
fastapi==0.111.0
requests>=2.28.2
pydantic_settings>=2.2.1
sqlalchemy>=2.0.28
psycopg2-binary
aiohttp
beanie
pyyaml
shapely
langchain
langchain-google-vertexai
twilio
haversine
